requirements.txt
streamlit
textblob
nltk
plotly
pandas
python-docx
pypdf
flake8
pytest
coverage